feat: Chapter4 폴더로부터 root 디렉토리로 이동, 및 파일 수정
