--- Chapter4/src/main/java/com/example/chapter4/domain/member/entity/mapping/MemberTerm.java
@@ -1,5 +1,0 @@
-package com.example.chapter4.domain.member.entity.mapping;
-
-public class MemberTerm {
-  
-}
