--- Chapter4/src/main/java/com/example/chapter4/domain/member/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.example.chapter4.domain.member.enums;
-
-public enum Gender {
-    MAN, WOMAN, NONE
-}
