+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 group = 'com.example'
 version = '0.0.1-SNAPSHOT'
-description = 'Chapter4'
+description = 'umc9th'
 
 java {
     toolchain {
