+++ src/main/java/com/example/umc9th/Umc9thApplication.java
@@ -1,4 +1,4 @@
-package com.example.chapter4;
+package com.example.umc9th;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 @EnableJpaAuditing
-public class Chapter4Application {
+public class Umc9thApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Chapter4Application.class, args);
+        SpringApplication.run(Umc9thApplication.class, args);
     }
 
 }
