+++ src/main/java/com/example/umc9th/domain/member/entity/mapping/MemberFood.java
@@ -1,7 +1,7 @@
-package com.example.chapter4.domain.member.entity.mapping;
+package com.example.umc9th.domain.member.entity.mapping;
 
-import com.example.chapter4.domain.member.entity.Food;
-import com.example.chapter4.domain.member.entity.Member;
+import com.example.umc9th.domain.member.entity.Food;
+import com.example.umc9th.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.*;
