+++ src/main/java/com/example/umc9th/domain/member/entity/mapping/MemberTerm.java
@@ -1,0 +1,5 @@
+package com.example.umc9th.domain.member.entity.mapping;
+
+public class MemberTerm {
+  
+}
