+++ src/main/java/com/example/umc9th/domain/member/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.example.umc9th.domain.member.enums;
+
+public enum Gender {
+    MAN, WOMAN, NONE
+}
