+++ src/main/java/com/example/umc9th/domain/member/enums/TermName.java
@@ -1,0 +1,5 @@
+package com.example.umc9th.domain.member.enums;
+
+public enum TermName {
+  
+}
