+++ src/test/java/com/example/umc9th/Umc9thApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.chapter4;
+package com.example.umc9th;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class Chapter4ApplicationTests {
+class Umc9thApplicationTests {
 
     @Test
     void contextLoads() {
